+++ alkahest-core/src/poly/resultant.rs
@@ -108,6 +108,7 @@ fn collect_vars_rec(expr: ExprId, pool: &ExprPool, out: &mut BTreeSet<ExprId>) {
         }
         ExprData::Predicate { args, .. } => args.clone(),
         ExprData::Forall { var, body } | ExprData::Exists { var, body } => vec![*var, *body],
+        ExprData::BigO(arg) => vec![*arg],
     });
     for child in children {
         collect_vars_rec(child, pool, out);